+++ samples-and-tests/excel/conf/application.conf
@@ -29,7 +29,7 @@ module.excel=../../
 # excel.template.root=app/views
 # Set the template root to any folder other than app/views if
 #     use this module on play version before 1.1
-excel.template.root=app/excel_tmpl
+# excel.template.root=app/excel_tmpl
 
 # i18n
 # ~~~~~
