+++ samples-and-tests/excel/conf/dependencies.yml
@@ -1,0 +1,2 @@
+require:
+    - play -> excel trunk
