+++ src/play.plugins
@@ -1,0 +1,1 @@
+200:play.modules.excel.Plugin
